refactor: return indexing to MediaManager, remove delegated tests
- Return IndexSeriesLibrary, ensureSeries, getOrCreateSeason, upsertEpisodeFile,
  matchEpisodeFile, modelEpisodeFile to MediaManager (series_index.go).
  These will be extracted into IndexerService in a separate task.
- Remove library dependency from SeriesService (still needed for DeleteSeries
  directory cleanup, but indexing methods no longer on SeriesService).
- Remove delegated tests from manager_test.go: TestListShowsInLibrary,
  TestBuildTVDetailResult, TestGetTVDetailByTMDBID,
  TestMediaManager_AddSeriesToLibrary (already covered by SeriesService tests)
- Delete tv_detail_test.go entirely (delegated, covered by SeriesService tests)
- Remove indexing tests from series_service_test.go (ensureSeries, IndexSeriesLibrary)
- Clean up unused imports

Net reduction: ~1350 lines of delegation testing and moved indexing code.
SeriesService now owns only series CRUD + TV detail/listing operations.

Part of #139
