+++ pkg/manager/manager_test.go
@@ -14,7 +14,6 @@ import (
 	"time"
 
 	"github.com/gkampitakis/go-snaps/snaps"
-	"github.com/go-jet/jet/v2/sqlite"
 	"github.com/kasuboski/mediaz/config"
 	downloadMock "github.com/kasuboski/mediaz/pkg/download/mocks"
 	mhttpMock "github.com/kasuboski/mediaz/pkg/http/mocks"
@@ -28,9 +27,7 @@ import (
 	storageMocks "github.com/kasuboski/mediaz/pkg/storage/mocks"
 	mediaSqlite "github.com/kasuboski/mediaz/pkg/storage/sqlite"
 	"github.com/kasuboski/mediaz/pkg/storage/sqlite/schema/gen/model"
-	"github.com/kasuboski/mediaz/pkg/storage/sqlite/schema/gen/table"
 	"github.com/kasuboski/mediaz/pkg/tmdb"
-	tmdbMocks "github.com/kasuboski/mediaz/pkg/tmdb/mocks"
 	"github.com/oapi-codegen/nullable"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -221,98 +218,6 @@ func TestListMoviesInLibrary(t *testing.T) {
 		movies, err := m.ListMoviesInLibrary(ctx)
 		assert.Error(t, err)
 		assert.Nil(t, movies)
-	})
-}
-
-func TestListShowsInLibrary(t *testing.T) {
-	ctx := context.Background()
-
-	t.Run("no shows in library", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		store := storageMocks.NewMockStorage(ctrl)
-		m := New(nil, nil, nil, store, nil, config.Manager{}, config.Config{})
-
-		store.EXPECT().ListSeries(ctx).Return([]*storage.Series{}, nil)
-
-		shows, err := m.ListShowsInLibrary(ctx)
-		require.NoError(t, err)
-		assert.Empty(t, shows)
-	})
-
-	t.Run("shows with metadata", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		store := storageMocks.NewMockStorage(ctrl)
-		m := New(nil, nil, nil, store, nil, config.Manager{}, config.Config{})
-
-		metadataID := int32(1)
-		path := "Show 1"
-		series := &storage.Series{
-			Series: model.Series{
-				SeriesMetadataID: &metadataID,
-				Path:             &path,
-			},
-			State: storage.SeriesStateDiscovered,
-		}
-
-		seriesMetadata := &model.SeriesMetadata{
-			ID:             1,
-			TmdbID:         321,
-			Title:          "Test Series",
-			Status:         "Continuing",
-			PosterPath:     ptr.To("poster.jpg"),
-			ExternalIds:    nil,
-			WatchProviders: nil,
-		}
-
-		store.EXPECT().ListSeries(ctx).Return([]*storage.Series{series}, nil)
-		store.EXPECT().GetSeriesMetadata(ctx, gomock.Any()).Return(seriesMetadata, nil)
-
-		shows, err := m.ListShowsInLibrary(ctx)
-		require.NoError(t, err)
-		require.Len(t, shows, 1)
-
-		expected := LibraryShow{
-			Path:       path,
-			TMDBID:     seriesMetadata.TmdbID,
-			Title:      seriesMetadata.Title,
-			PosterPath: *seriesMetadata.PosterPath,
-		}
-		assert.Equal(t, expected.Path, shows[0].Path)
-		assert.Equal(t, expected.TMDBID, shows[0].TMDBID)
-		assert.Equal(t, expected.Title, shows[0].Title)
-		assert.Equal(t, expected.PosterPath, shows[0].PosterPath)
-	})
-
-	t.Run("shows without metadata are filtered out", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		store := storageMocks.NewMockStorage(ctrl)
-		m := New(nil, nil, nil, store, nil, config.Manager{}, config.Config{})
-
-		path := "Show 1"
-		series := &storage.Series{
-			Series: model.Series{
-				Path: &path,
-			},
-			State: storage.SeriesStateDiscovered,
-		}
-
-		store.EXPECT().ListSeries(ctx).Return([]*storage.Series{series}, nil)
-
-		shows, err := m.ListShowsInLibrary(ctx)
-		require.NoError(t, err)
-		assert.Empty(t, shows, "series without metadata should not be included in library")
-	})
-
-	t.Run("error listing series", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		store := storageMocks.NewMockStorage(ctrl)
-		m := New(nil, nil, nil, store, nil, config.Manager{}, config.Config{})
-
-		store.EXPECT().ListSeries(ctx).Return(nil, errors.New("db error"))
-
-		shows, err := m.ListShowsInLibrary(ctx)
-		assert.Error(t, err)
-		assert.Nil(t, shows)
 	})
 }
 
@@ -888,467 +793,6 @@ func TestGetMovieDetailByTMDBID(t *testing.T) {
 	})
 }
 
-func TestBuildTVDetailResult(t *testing.T) {
-	t.Run("builds complete TV detail result", func(t *testing.T) {
-		m := New(nil, nil, nil, nil, nil, config.Manager{}, config.Config{})
-
-		// Mock series metadata
-		firstAirDate := time.Date(2023, 1, 15, 0, 0, 0, 0, time.UTC)
-		lastAirDate := time.Date(2023, 12, 15, 0, 0, 0, 0, time.UTC)
-		metadata := &model.SeriesMetadata{
-			ID:             1,
-			TmdbID:         123,
-			Title:          "Test Series",
-			Status:         "Continuing",
-			Overview:       ptr.To("Test series overview"),
-			FirstAirDate:   &firstAirDate,
-			LastAirDate:    &lastAirDate,
-			SeasonCount:    2,
-			EpisodeCount:   20,
-			ExternalIds:    nil,
-			WatchProviders: nil,
-		}
-
-		// Mock TMDB details response
-		details := &tmdb.SeriesDetailsResponse{
-			PosterPath:   "poster.jpg",
-			BackdropPath: "backdrop.jpg",
-			Adult:        true,
-			Popularity:   8.5,
-			Networks: []tmdb.Network{
-				{Name: "HBO"},
-				{Name: "Netflix"},
-			},
-			Genres: []tmdb.Genre{
-				{Name: "Drama"},
-				{Name: "Thriller"},
-			},
-		}
-
-		// Mock library series - in library and monitored
-		path := "/tv/test-series"
-		qualityProfileID := int32(2)
-		series := &storage.Series{
-			Series: model.Series{
-				ID:               1,
-				SeriesMetadataID: ptr.To(int32(1)),
-				Path:             &path,
-				QualityProfileID: qualityProfileID,
-				Monitored:        1,
-			},
-			State: storage.SeriesStateDiscovered,
-		}
-
-		result := m.buildTVDetailResult(metadata, details, series, []SeasonResult{})
-
-		assert.Equal(t, int32(123), result.TMDBID)
-		assert.Equal(t, "Test Series", result.Title)
-		assert.Equal(t, "Test series overview", result.Overview)
-		assert.Equal(t, "poster.jpg", result.PosterPath)
-		assert.Equal(t, ptr.To("backdrop.jpg"), result.BackdropPath)
-		assert.Equal(t, ptr.To("2023-01-15"), result.FirstAirDate)
-		assert.Equal(t, ptr.To("2023-12-15"), result.LastAirDate)
-		assert.Equal(t, int32(2), result.SeasonCount)
-		assert.Equal(t, int32(20), result.EpisodeCount)
-		require.Len(t, result.Networks, 2)
-		assert.Equal(t, "HBO", result.Networks[0].Name)
-		assert.Equal(t, "Netflix", result.Networks[1].Name)
-		assert.Equal(t, []string{"Drama", "Thriller"}, result.Genres)
-		assert.Equal(t, ptr.To(true), result.Adult)
-		pop := float64(8.5)
-		assert.Equal(t, &pop, result.Popularity)
-		assert.Equal(t, string(storage.SeriesStateDiscovered), result.LibraryStatus)
-		assert.Equal(t, &path, result.Path)
-		assert.Equal(t, &qualityProfileID, result.QualityProfileID)
-		monitored := true
-		assert.Equal(t, &monitored, result.Monitored)
-	})
-
-	t.Run("builds TV detail result not in library", func(t *testing.T) {
-		m := New(nil, nil, nil, nil, nil, config.Manager{}, config.Config{})
-
-		metadata := &model.SeriesMetadata{
-			ID:             1,
-			TmdbID:         123,
-			Title:          "Test Series",
-			Status:         "Continuing",
-			SeasonCount:    1,
-			EpisodeCount:   10,
-			ExternalIds:    nil,
-			WatchProviders: nil,
-		}
-
-		details := &tmdb.SeriesDetailsResponse{
-			PosterPath:   "poster.jpg",
-			BackdropPath: "backdrop.jpg",
-		}
-
-		// No series in library (nil)
-		result := m.buildTVDetailResult(metadata, details, nil, []SeasonResult{})
-
-		assert.Equal(t, int32(123), result.TMDBID)
-		assert.Equal(t, "Test Series", result.Title)
-		assert.Equal(t, "Not In Library", result.LibraryStatus)
-		assert.Nil(t, result.Path)
-		assert.Nil(t, result.QualityProfileID)
-		assert.Nil(t, result.Monitored)
-	})
-
-	t.Run("builds TV detail result with minimal data", func(t *testing.T) {
-		m := New(nil, nil, nil, nil, nil, config.Manager{}, config.Config{})
-
-		metadata := &model.SeriesMetadata{
-			ID:             1,
-			TmdbID:         123,
-			Title:          "Test Series",
-			Status:         "Continuing",
-			SeasonCount:    1,
-			EpisodeCount:   5,
-			ExternalIds:    nil,
-			WatchProviders: nil,
-		}
-
-		details := &tmdb.SeriesDetailsResponse{
-			PosterPath: "poster.jpg",
-			// No backdrop, networks, genres, etc.
-		}
-
-		result := m.buildTVDetailResult(metadata, details, nil, []SeasonResult{})
-
-		assert.Equal(t, int32(123), result.TMDBID)
-		assert.Equal(t, "Test Series", result.Title)
-		assert.Equal(t, "poster.jpg", result.PosterPath)
-		assert.Nil(t, result.BackdropPath)
-		assert.Nil(t, result.FirstAirDate)
-		assert.Nil(t, result.LastAirDate)
-		assert.Empty(t, result.Networks)
-		assert.Empty(t, result.Genres)
-		assert.Nil(t, result.Adult)
-		assert.Nil(t, result.Popularity)
-		assert.Equal(t, "Not In Library", result.LibraryStatus)
-	})
-
-	t.Run("builds TV detail result with unmonitored series", func(t *testing.T) {
-		m := New(nil, nil, nil, nil, nil, config.Manager{}, config.Config{})
-
-		metadata := &model.SeriesMetadata{
-			ID:             1,
-			TmdbID:         123,
-			Title:          "Test Series",
-			Status:         "Continuing",
-			SeasonCount:    1,
-			EpisodeCount:   5,
-			ExternalIds:    nil,
-			WatchProviders: nil,
-		}
-
-		details := &tmdb.SeriesDetailsResponse{
-			PosterPath: "poster.jpg",
-		}
-
-		// Unmonitored series
-		series := &storage.Series{
-			Series: model.Series{
-				ID:               1,
-				SeriesMetadataID: ptr.To(int32(1)),
-				QualityProfileID: 1,
-				Monitored:        0, // Unmonitored
-			},
-			State: storage.SeriesStateMissing,
-		}
-
-		result := m.buildTVDetailResult(metadata, details, series, []SeasonResult{})
-
-		monitored := false
-		assert.Equal(t, &monitored, result.Monitored)
-		assert.Equal(t, string(storage.SeriesStateMissing), result.LibraryStatus)
-	})
-}
-
-func TestGetTVDetailByTMDBID(t *testing.T) {
-	ctx := context.Background()
-
-	t.Run("success - TV show not in library", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		store := storageMocks.NewMockStorage(ctrl)
-		tmdbMock := tmdbMocks.NewMockITmdb(ctrl)
-
-		m := New(tmdbMock, nil, nil, store, nil, config.Manager{}, config.Config{})
-
-		firstAirDate := time.Date(2023, 1, 15, 0, 0, 0, 0, time.UTC)
-		externalIDsJSON := `{"imdb_id":"tt7654321","tvdb_id":54321}`
-		watchProvidersJSON := `{"US":{"flatrate":[{"provider_id":8,"provider_name":"Netflix","logo_path":"/net.png"}]}}`
-		metadata := &model.SeriesMetadata{
-			ID:             1,
-			TmdbID:         123,
-			Title:          "Test Series",
-			Overview:       ptr.To("Test series overview"),
-			FirstAirDate:   &firstAirDate,
-			SeasonCount:    2,
-			EpisodeCount:   20,
-			Status:         "Continuing",
-			ExternalIds:    &externalIDsJSON,
-			WatchProviders: &watchProvidersJSON,
-		}
-
-		// Mock series details response will be returned via HTTP response
-
-		// Mock GetSeriesMetadata call
-		store.EXPECT().GetSeriesMetadata(ctx, gomock.Any()).Return(metadata, nil)
-
-		// Mock TvSeriesDetails call
-		responseBody := `{"poster_path":"poster.jpg","backdrop_path":"backdrop.jpg","networks":[{"name":"HBO"}],"genres":[{"name":"Drama"}]}`
-		resp := &http.Response{
-			StatusCode: 200,
-			Body:       io.NopCloser(bytes.NewBufferString(responseBody)),
-		}
-		tmdbMock.EXPECT().TvSeriesDetails(ctx, int32(123), nil).Return(resp, nil)
-
-		// Mock GetSeries call - not found
-		store.EXPECT().GetSeries(ctx, gomock.Any()).Return(nil, storage.ErrNotFound)
-
-		result, err := m.GetTVDetailByTMDBID(ctx, 123)
-		require.NoError(t, err)
-		require.NotNil(t, result)
-
-		assert.Equal(t, int32(123), result.TMDBID)
-		assert.Equal(t, "Test Series", result.Title)
-		assert.Equal(t, "Test series overview", result.Overview)
-		assert.Equal(t, "poster.jpg", result.PosterPath)
-		assert.Equal(t, ptr.To("backdrop.jpg"), result.BackdropPath)
-		assert.Equal(t, ptr.To("2023-01-15"), result.FirstAirDate)
-		assert.Equal(t, int32(2), result.SeasonCount)
-		assert.Equal(t, int32(20), result.EpisodeCount)
-		require.Len(t, result.Networks, 1)
-		assert.Equal(t, "HBO", result.Networks[0].Name)
-		assert.Equal(t, []string{"Drama"}, result.Genres)
-		assert.Equal(t, "Not In Library", result.LibraryStatus)
-		assert.Nil(t, result.Path)
-		assert.Nil(t, result.QualityProfileID)
-		assert.Nil(t, result.Monitored)
-		// Check external IDs from stored data
-		require.NotNil(t, result.ExternalIDs)
-		assert.Equal(t, ptr.To("tt7654321"), result.ExternalIDs.ImdbID)
-		tvdbID := 54321
-		assert.Equal(t, &tvdbID, result.ExternalIDs.TvdbID)
-		// Check watch providers from stored data
-		require.Len(t, result.WatchProviders, 1)
-		assert.Equal(t, 8, result.WatchProviders[0].ProviderID)
-		assert.Equal(t, "Netflix", result.WatchProviders[0].Name)
-		assert.Equal(t, ptr.To("/net.png"), result.WatchProviders[0].LogoPath)
-	})
-
-	t.Run("success - TV show in library", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		store := storageMocks.NewMockStorage(ctrl)
-		tmdbMock := tmdbMocks.NewMockITmdb(ctrl)
-
-		m := New(tmdbMock, nil, nil, store, nil, config.Manager{}, config.Config{})
-
-		emptyExternalIDsJSON := `{"imdb_id":null,"tvdb_id":null}`
-		emptyWatchProvidersJSON := `{"US":{"flatrate":[]}}`
-		metadata := &model.SeriesMetadata{
-			ID:             1,
-			TmdbID:         123,
-			Title:          "Test Series",
-			SeasonCount:    1,
-			EpisodeCount:   10,
-			Status:         "Continuing",
-			ExternalIds:    &emptyExternalIDsJSON,
-			WatchProviders: &emptyWatchProvidersJSON,
-		}
-
-		// Mock series details response will be returned via HTTP response
-
-		path := "/tv/test-series"
-		qualityProfileID := int32(2)
-		series := &storage.Series{
-			Series: model.Series{
-				ID:               1,
-				SeriesMetadataID: ptr.To(int32(1)),
-				Path:             &path,
-				QualityProfileID: qualityProfileID,
-				Monitored:        1,
-			},
-			State: storage.SeriesStateDiscovered,
-		}
-
-		store.EXPECT().GetSeriesMetadata(ctx, gomock.Any()).Return(metadata, nil)
-
-		responseBody := `{"poster_path":"poster.jpg","backdrop_path":"backdrop.jpg"}`
-		resp := &http.Response{
-			StatusCode: 200,
-			Body:       io.NopCloser(bytes.NewBufferString(responseBody)),
-		}
-		tmdbMock.EXPECT().TvSeriesDetails(ctx, int32(123), nil).Return(resp, nil)
-
-		store.EXPECT().GetSeries(ctx, gomock.Any()).Return(series, nil)
-
-		// Mock calls for seasons and episodes (empty list since this test doesn't focus on seasons)
-		store.EXPECT().ListSeasons(ctx, gomock.Any()).Return([]*storage.Season{}, nil)
-
-		result, err := m.GetTVDetailByTMDBID(ctx, 123)
-		require.NoError(t, err)
-		require.NotNil(t, result)
-
-		assert.Equal(t, int32(123), result.TMDBID)
-		assert.Equal(t, "Test Series", result.Title)
-		assert.Equal(t, string(storage.SeriesStateDiscovered), result.LibraryStatus)
-		assert.Equal(t, &path, result.Path)
-		assert.Equal(t, &qualityProfileID, result.QualityProfileID)
-		monitored := true
-		assert.Equal(t, &monitored, result.Monitored)
-	})
-
-	t.Run("error - GetSeriesMetadata fails", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		store := storageMocks.NewMockStorage(ctrl)
-		tmdbMock := tmdbMocks.NewMockITmdb(ctrl)
-
-		m := New(tmdbMock, nil, nil, store, nil, config.Manager{}, config.Config{})
-
-		expectedErr := errors.New("metadata fetch error")
-		store.EXPECT().GetSeriesMetadata(ctx, gomock.Any()).Return(nil, expectedErr)
-
-		result, err := m.GetTVDetailByTMDBID(ctx, 123)
-		assert.Error(t, err)
-		assert.Equal(t, expectedErr, err)
-		assert.Nil(t, result)
-	})
-
-	t.Run("error - TMDB API call fails", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		store := storageMocks.NewMockStorage(ctrl)
-		tmdbMock := tmdbMocks.NewMockITmdb(ctrl)
-
-		m := New(tmdbMock, nil, nil, store, nil, config.Manager{}, config.Config{})
-
-		metadata := &model.SeriesMetadata{
-			ID:             1,
-			TmdbID:         123,
-			Title:          "Test Series",
-			Status:         "Continuing",
-			ExternalIds:    nil,
-			WatchProviders: nil,
-		}
-
-		store.EXPECT().GetSeriesMetadata(ctx, gomock.Any()).Return(metadata, nil)
-
-		expectedErr := errors.New("TMDB API error")
-		tmdbMock.EXPECT().TvSeriesDetails(ctx, int32(123), nil).Return(nil, expectedErr)
-
-		result, err := m.GetTVDetailByTMDBID(ctx, 123)
-		assert.Error(t, err)
-		assert.Equal(t, expectedErr, err)
-		assert.Nil(t, result)
-	})
-
-	t.Run("success - storage error non-NotFound is logged but not returned", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		store := storageMocks.NewMockStorage(ctrl)
-		tmdbMock := tmdbMocks.NewMockITmdb(ctrl)
-
-		m := New(tmdbMock, nil, nil, store, nil, config.Manager{}, config.Config{})
-
-		metadata := &model.SeriesMetadata{
-			ID:             1,
-			TmdbID:         123,
-			Title:          "Test Series",
-			Status:         "Continuing",
-			ExternalIds:    nil,
-			WatchProviders: nil,
-		}
-
-		// Mock series details response will be returned via HTTP response
-
-		store.EXPECT().GetSeriesMetadata(ctx, gomock.Any()).Return(metadata, nil)
-
-		responseBody := `{"poster_path":"poster.jpg"}`
-		resp := &http.Response{
-			StatusCode: 200,
-			Body:       io.NopCloser(bytes.NewBufferString(responseBody)),
-		}
-		tmdbMock.EXPECT().TvSeriesDetails(ctx, int32(123), nil).Return(resp, nil)
-
-		dbErr := errors.New("database connection error")
-		store.EXPECT().GetSeries(ctx, gomock.Any()).Return(nil, dbErr)
-
-		result, err := m.GetTVDetailByTMDBID(ctx, 123)
-		require.NoError(t, err)
-		require.NotNil(t, result)
-
-		assert.Equal(t, int32(123), result.TMDBID)
-		assert.Equal(t, "Not In Library", result.LibraryStatus)
-		assert.Nil(t, result.Path)
-		assert.Nil(t, result.QualityProfileID)
-		assert.Nil(t, result.Monitored)
-	})
-
-	t.Run("error - TMDB returns empty FirstAirDate", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		store := storageMocks.NewMockStorage(ctrl)
-		tmdbMock := tmdbMocks.NewMockITmdb(ctrl)
-
-		m := New(tmdbMock, nil, nil, store, nil, config.Manager{}, config.Config{})
-
-		// This should trigger the creation of metadata which calls FromSeriesDetails with empty FirstAirDate
-		store.EXPECT().GetSeriesMetadata(ctx, gomock.Any()).Return(nil, storage.ErrNotFound)
-
-		// Mock GetSeriesDetails to return series with empty FirstAirDate
-		tmdbMock.EXPECT().GetSeriesDetails(ctx, 123).Return(&tmdb.SeriesDetails{
-			ID:               123,
-			Name:             "Test Series",
-			FirstAirDate:     "", // Empty date that causes the parsing error
-			NumberOfSeasons:  1,
-			NumberOfEpisodes: 10,
-			Seasons:          []tmdb.Season{}, // Empty to avoid creating season metadata
-		}, nil)
-
-		// Mock external IDs and watch providers calls during metadata creation
-		extIDsResp := &http.Response{StatusCode: 200, Body: io.NopCloser(bytes.NewBufferString(`{"imdb_id":null,"tvdb_id":null}`))}
-		tmdbMock.EXPECT().TvSeriesExternalIds(ctx, int32(123)).Return(extIDsResp, nil)
-		wpResp := &http.Response{StatusCode: 200, Body: io.NopCloser(bytes.NewBufferString(`{"results":{"US":{"flatrate":[]}}}`))}
-		tmdbMock.EXPECT().TvSeriesWatchProviders(ctx, int32(123)).Return(wpResp, nil)
-
-		// Mock CreateSeriesMetadata call
-		store.EXPECT().CreateSeriesMetadata(ctx, gomock.Any()).Return(int64(1), nil)
-
-		// Mock GetSeriesMetadata call after creation
-		createdMetadata := &model.SeriesMetadata{
-			ID:             1,
-			TmdbID:         123,
-			Title:          "Test Series",
-			Status:         "Continuing",
-			FirstAirDate:   nil, // Should be nil due to empty date
-			SeasonCount:    1,
-			EpisodeCount:   10,
-			ExternalIds:    nil,
-			WatchProviders: nil,
-		}
-		store.EXPECT().GetSeriesMetadata(ctx, gomock.Any()).Return(createdMetadata, nil)
-
-		// Mock TvSeriesDetails call for the main flow
-		responseBody := `{"poster_path":"poster.jpg"}`
-		resp := &http.Response{
-			StatusCode: 200,
-			Body:       io.NopCloser(bytes.NewBufferString(responseBody)),
-		}
-		tmdbMock.EXPECT().TvSeriesDetails(ctx, int32(123), nil).Return(resp, nil)
-
-		// Mock GetSeries call - not found
-		store.EXPECT().GetSeries(ctx, gomock.Any()).Return(nil, storage.ErrNotFound)
-
-		result, err := m.GetTVDetailByTMDBID(ctx, 123)
-		require.NoError(t, err, "Should handle empty FirstAirDate gracefully")
-		require.NotNil(t, result)
-
-		assert.Equal(t, int32(123), result.TMDBID)
-		assert.Equal(t, "Test Series", result.Title)
-		assert.Nil(t, result.FirstAirDate) // Should be nil when date is empty
-		assert.Equal(t, "Not In Library", result.LibraryStatus)
-	})
-}
-
 func TestParseTMDBDate(t *testing.T) {
 	t.Run("parses valid date", func(t *testing.T) {
 		result, err := parseTMDBDate("2023-01-15")
@@ -1369,226 +813,6 @@ func TestParseTMDBDate(t *testing.T) {
 		result, err := parseTMDBDate("invalid-date")
 		assert.Error(t, err)
 		assert.Nil(t, result)
-	})
-}
-
-func TestMediaManager_AddSeriesToLibrary(t *testing.T) {
-	t.Run("error getting quality profile", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		ctx := context.Background()
-
-		store := storageMocks.NewMockStorage(ctrl)
-		store.EXPECT().GetQualityProfile(gomock.Any(), int64(1)).Return(storage.QualityProfile{}, errors.New("expected testing error"))
-
-		m := New(nil, nil, nil, store, nil, config.Manager{}, config.Config{})
-		require.NotNil(t, m)
-
-		req := AddSeriesRequest{
-			TMDBID:           1234,
-			QualityProfileID: 1,
-		}
-
-		_, err := m.AddSeriesToLibrary(ctx, req)
-		assert.Error(t, err)
-		assert.Equal(t, "expected testing error", err.Error())
-	})
-
-	t.Run("error getting series metadata from tdmb", func(t *testing.T) {
-		ctx := context.Background()
-		ctrl := gomock.NewController(t)
-
-		store := newStore(t, ctx)
-		store.CreateQualityProfile(ctx, model.QualityProfile{
-			ID:   1,
-			Name: "test-profile",
-		})
-
-		tmdbMock := tmdbMocks.NewMockITmdb(ctrl)
-		tmdbMock.EXPECT().GetSeriesDetails(gomock.Any(), gomock.Any()).Return(nil, errors.New("expected testing error"))
-
-		m := New(tmdbMock, nil, nil, store, nil, config.Manager{}, config.Config{})
-		require.NotNil(t, m)
-
-		req := AddSeriesRequest{
-			TMDBID:           1234,
-			QualityProfileID: 1,
-		}
-
-		_, err := m.AddSeriesToLibrary(ctx, req)
-		assert.Error(t, err)
-		assert.Equal(t, "expected testing error", err.Error())
-	})
-
-	t.Run("series metadata exists in db - series exists", func(t *testing.T) {
-		ctx := context.Background()
-		ctrl := gomock.NewController(t)
-
-		store := newStore(t, ctx)
-		_, err := store.CreateQualityProfile(ctx, model.QualityProfile{
-			ID:   1,
-			Name: "test-profile",
-		})
-		require.NoError(t, err)
-
-		metadataID, err := store.CreateSeriesMetadata(ctx, model.SeriesMetadata{
-			ID:             1,
-			TmdbID:         1234,
-			Title:          "Test Series",
-			SeasonCount:    1,
-			EpisodeCount:   1,
-			Status:         "Continuing",
-			ExternalIds:    nil, // Optional field
-			WatchProviders: nil, // Optional field
-		})
-		require.NoError(t, err)
-
-		seriesID, err := store.CreateSeries(ctx, storage.Series{
-			Series: model.Series{
-				ID:               1,
-				SeriesMetadataID: ptr.To(int32(metadataID)),
-				Monitored:        1,
-				QualityProfileID: 1,
-				Added:            ptr.To(time.Now()),
-			},
-		}, storage.SeriesStateMissing)
-		require.NoError(t, err)
-
-		tmdbMock := tmdbMocks.NewMockITmdb(ctrl)
-
-		m := New(tmdbMock, nil, nil, store, nil, config.Manager{}, config.Config{})
-		require.NotNil(t, m)
-
-		req := AddSeriesRequest{
-			TMDBID:           1234,
-			QualityProfileID: 1,
-		}
-
-		series, err := m.AddSeriesToLibrary(ctx, req)
-		require.NoError(t, err)
-
-		assert.Equal(t, series.ID, int32(seriesID))
-		assert.Equal(t, series.SeriesMetadataID, ptr.To(int32(metadataID)))
-		assert.Equal(t, series.Monitored, int32(1))
-		assert.Equal(t, series.QualityProfileID, int32(1))
-		assert.Equal(t, storage.SeriesStateMissing, series.State)
-	})
-
-	t.Run("series metadata exists in db - series does not exist - unreleased", func(t *testing.T) {
-		ctx := context.Background()
-		ctrl := gomock.NewController(t)
-
-		store := newStore(t, ctx)
-		_, err := store.CreateQualityProfile(ctx, model.QualityProfile{
-			ID:   1,
-			Name: "test-profile",
-		})
-		require.NoError(t, err)
-
-		metadataID, err := store.CreateSeriesMetadata(ctx, model.SeriesMetadata{
-			ID:             1,
-			TmdbID:         1234,
-			Title:          "Test Series",
-			SeasonCount:    1,
-			EpisodeCount:   1,
-			Status:         "Continuing",
-			FirstAirDate:   ptr.To(time.Now().Add(time.Hour * 24 * 7)),
-			ExternalIds:    nil, // Optional field
-			WatchProviders: nil, // Optional field
-		})
-		require.NoError(t, err)
-
-		tmdbMock := tmdbMocks.NewMockITmdb(ctrl)
-
-		m := New(tmdbMock, nil, nil, store, nil, config.Manager{}, config.Config{})
-		require.NotNil(t, m)
-
-		req := AddSeriesRequest{
-			TMDBID:           1234,
-			QualityProfileID: 1,
-		}
-
-		series, err := m.AddSeriesToLibrary(ctx, req)
-		require.NoError(t, err)
-
-		assert.Equal(t, series.ID, int32(1))
-		assert.Equal(t, series.SeriesMetadataID, ptr.To(int32(metadataID)))
-		assert.Equal(t, series.Monitored, int32(1))
-		assert.Equal(t, series.QualityProfileID, int32(1))
-		assert.Equal(t, storage.SeriesStateUnreleased, series.State)
-	})
-
-	t.Run("series metadata exists in db - series does not exist - released", func(t *testing.T) {
-		ctx := context.Background()
-		ctrl := gomock.NewController(t)
-
-		store := newStore(t, ctx)
-		_, err := store.CreateQualityProfile(ctx, model.QualityProfile{
-			ID:   1,
-			Name: "test-profile",
-		})
-		require.NoError(t, err)
-
-		tmdbMock := tmdbMocks.NewMockITmdb(ctrl)
-		details := &tmdb.SeriesDetails{
-			ID:               1,
-			Name:             "test series",
-			FirstAirDate:     time.Now().Add(-time.Hour * 2).Format(tmdb.ReleaseDateFormat),
-			NumberOfSeasons:  1,
-			NumberOfEpisodes: 1,
-			Seasons: []tmdb.Season{
-				{
-					ID:           1,
-					Name:         "test season",
-					SeasonNumber: 1,
-					Episodes: []tmdb.Episode{
-						{
-							ID:            1,
-							Name:          "test episode",
-							EpisodeNumber: 1,
-						},
-					},
-				},
-			},
-		}
-
-		tmdbMock.EXPECT().GetSeriesDetails(gomock.Any(), gomock.Any()).Return(details, nil)
-
-		// Mock external IDs and watch providers calls during metadata creation
-		extIDsResp := &http.Response{StatusCode: 200, Body: io.NopCloser(bytes.NewBufferString(`{"imdb_id":null,"tvdb_id":null}`))}
-		tmdbMock.EXPECT().TvSeriesExternalIds(gomock.Any(), int32(1234)).Return(extIDsResp, nil)
-		wpResp := &http.Response{StatusCode: 200, Body: io.NopCloser(bytes.NewBufferString(`{"results":{"US":{"flatrate":[]}}}`))}
-		tmdbMock.EXPECT().TvSeriesWatchProviders(gomock.Any(), int32(1234)).Return(wpResp, nil)
-
-		m := New(tmdbMock, nil, nil, store, nil, config.Manager{}, config.Config{})
-		require.NotNil(t, m)
-
-		req := AddSeriesRequest{
-			TMDBID:           1234,
-			QualityProfileID: 1,
-		}
-
-		series, err := m.AddSeriesToLibrary(ctx, req)
-		require.NoError(t, err)
-
-		assert.Equal(t, int32(1), series.ID)
-		assert.Equal(t, ptr.To(int32(1)), series.SeriesMetadataID)
-		assert.Equal(t, int32(1), series.Monitored)
-		assert.Equal(t, int32(1), series.QualityProfileID)
-		assert.Equal(t, storage.SeriesStateMissing, series.State)
-
-		seasons, err := m.seriesStorage.ListSeasons(ctx, table.Season.SeriesID.EQ(sqlite.Int32(series.ID)))
-		assert.Nil(t, err)
-		require.Len(t, seasons, 1)
-		assert.Equal(t, int32(series.ID), seasons[0].SeriesID)
-		assert.Equal(t, seasons[0].State, storage.SeasonStateMissing)
-		assert.Equal(t, int32(1), seasons[0].Monitored)
-
-		episodes, err := m.seriesStorage.ListEpisodes(ctx, table.Episode.SeasonID.EQ(sqlite.Int32(seasons[0].ID)))
-		assert.Nil(t, err)
-		require.Len(t, episodes, 1)
-		assert.Equal(t, int32(1), episodes[0].EpisodeNumber)
-		assert.Equal(t, int32(1), episodes[0].Monitored)
-		assert.Equal(t, storage.EpisodeStateMissing, episodes[0].State)
 	})
 }
